% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_image.R
\name{gdal_image}
\alias{gdal_image}
\alias{gdal_raster}
\alias{gdal_map}
\alias{gdal_raster_data}
\title{Title}
\usage{
gdal_image(dsn = VirtualEarthSatellite)

gdal_raster(dsn = AWSElevation)

gdal_map(dsn = OSMMap)

gdal_raster_data(dsn = AWSElevation)
}
\arguments{
\item{dsn}{}
}
\value{

}
\description{
Title
}
\examples{
gdalio_set_default_grid(list(extent = c(-1, 1, -1, 1) * 500,
dimension = dev.size("px") * c(1, 1)/2,
projection = "+proj=laea +lon_0=153.02 +lat_0=-27.47"))
par(mfrow = c(2, 2))
plot(gdal_map())
plot(gdal_raster())
plot(gdal_image())
plot(gdal_map(VirtualEarthStreet))
#gdal_raster_data()
}
