% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{gdalio-example-files}
\alias{gdalio-example-files}
\alias{gdalio_eg_raster}
\alias{gdalio_eg_image}
\title{Example data files}
\usage{
gdalio_eg_raster()

gdalio_eg_image()
}
\description{
Return a path to a data raster file, for easy access to examples.
}
\details{
This raster has sea surface temperature values in Celsius, with no colour palette defined.
}
\examples{
gdalio_eg_raster()
gdalio_eg_image()
}
