% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalio.R
\name{gdalio_data}
\alias{gdalio_data}
\title{Title}
\usage{
gdalio_data(dsn, ...)
}
\arguments{
\item{dsn}{character string, raster source understood by GDAL}

\item{...}{arguments passed to 'vapour::vapour_warp_raster'}
}
\value{
list of numeric vectors
}
\description{
Title
}
\examples{
\dontrun{
 f <- system.file("extdata/sst.tif", package = "vapour", mustWork = TRUE)

 g <- list(extent = c(130, 160, -60, -30), dimension = c(180, 100),
    projection = "+proj=longlat")
 gdalio_set_default_grid(g)
 v <- gdalio_data(f)
 image(seq(130, 160, length.out = 181), seq(-60, -30, length.out = 101),
    matrix(v[[1]], g$dimension[1])[,g$dimension[2]:1], asp = 1.5)
}
}
