% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{gdalio_format_source}
\alias{gdalio_format_source}
\title{Print the code to source format-specific functions}
\usage{
gdalio_format_source()
}
\description{
You can 'source()' the file path displayed by this function to define package-specific formats for the gdalio data.
}
\details{
Running the code in the file path displayed by this function will load functions for terra, stars, raster, and spatstat.
}
\examples{
\dontrun{
source(gdalio_format_source())
}
}
