% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_grid.R
\name{gdalio_local_grid}
\alias{gdalio_local_grid}
\title{Generate a local grid}
\usage{
gdalio_local_grid(
  x = 147,
  y = -42,
  buffer = 2500000,
  family = "laea",
  dim = if (dev.cur() == 1) {
     c(512, 512)
 } else {
     dev.size("px")
 }
)
}
\arguments{
\item{x}{longitude}

\item{y}{latitude}

\item{buffer}{width either side of x, y}

\item{family}{projection family (as per PROJ strings)}

\item{dim}{size of grid nx, ny}
}
\value{
list appropriate for \code{\link[=gdalio_set_default_grid]{gdalio_set_default_grid()}}
}
\description{
Generate a default local grid to use for subsequent data reads.
}
\details{
All arguments have default values.
}
\examples{
gdalio_local_grid()
gdalio_local_grid(family = "stere")
}
