% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_grid.R
\name{gdalio_set_default_grid}
\alias{gdalio_set_default_grid}
\title{Title}
\usage{
gdalio_set_default_grid(x)
}
\arguments{
\item{x}{grid specification, a list with '$extent, $dimension, $projection` or a spatial grid object see Details}
}
\value{
nothing, used to set a default grid available globally
}
\description{
Input may be a list with \code{extent} \verb{$dimension}, \verb{$projection}, which is \code{c(xmin, xmax, ymin, ymax)},
\code{c(ncol, nrow)}, and string (accepted by GDAL as a projection input). Alternatively, use
a raster, stars, or terra object. Only simple cases of stars will work (regular grids with positive x, negative y transforms).
}
\examples{
gdalio_set_default_grid(list(extent = c(-1000, 1000, -2000, 2000),
   dimension = c(100, 200), projection = "+proj=longlat"))
gdalio_set_default_grid()
}
