% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vrt.R
\name{vrt}
\alias{vrt}
\alias{vrt.default}
\title{Fix missing raster metadata}
\usage{
vrt(x, extent = NULL, projection = NULL)

\method{vrt}{default}(x, extent = NULL, projection = NULL)
}
\arguments{
\item{x}{character string, file, url, GDAL dsn}

\item{extent}{numeric 'c(xmin, xmax, ymin, ymax)'}

\item{projection}{character wkt, proj, epsg code}
}
\value{
lightly classed character vector, with "vrt_simple", "character"
}
\description{
Simple metadata augmentation for raster sources.
}
\details{
Simple function to add either or both of a raster source extent and projection string.

The attributes from 'extent' as 'source_extent' and/or 'projection' as 'source_projection' are passed directly down to GDAL, via
the \code{\link[=gdalio_data]{gdalio_data()}} function, which hands them on to \code{\link[vapour:vapour_warp_raster]{vapour::vapour_warp_raster()}} arguments 'source_extent' and 'source_wkt'
respectively.
}
\examples{
vrt("myfile.nc")
vrt("myfile.nc", extent = c(-180, 180, -90, 90))
str(vrt("myfile.nc", extent = c(-180, 180, -90, 90), projection = 4326))
}
