% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawio.R
\name{drawio_data}
\alias{drawio_data}
\title{Read GDAL drawing geometry and fields data}
\usage{
drawio_data(dsn, ...)
}
\arguments{
\item{dsn}{character string, drawing data source understood by GDAL}

\item{...}{arguments passed to 'vapour::vapour_read_attributes()'}
}
\value{
list of numeric vectors
}
\description{
Read GDAL drawing geometry and fields data
}
