% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalio.R
\name{gdalio_format_source}
\alias{gdalio_format_source}
\title{Print the code to source format-specific functions}
\usage{
gdalio_format_source()
}
\description{
You can run the code displayed by this function to define package-specific formats for the gdalio data.
}
\details{
Currently running the code displayed by this function will load functions for terra, stars, raster, and spatstat.
}
\examples{
gdalio_format_source()
}
