% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_grid.R
\name{gdalio_set_default_grid}
\alias{gdalio_set_default_grid}
\title{Title}
\usage{
gdalio_set_default_grid(x)
}
\arguments{
\item{x}{optional grid specification}
}
\value{
nothing, used to set a default grid available globally
}
\description{
Title
}
\examples{
gdalio_set_default_grid(list(extent = c(-1000, 1000, -2000, 2000),
   dimension = c(100, 200), projection = "+proj=longlat"))
gdalio_set_default_grid()
}
